Refactor conversion. Store references when needed.
