+++ src/convert/convert.hpp
@@ -1,0 +1,24 @@
+#ifndef CONVERT_H
+#define CONVERT_H
+
+#include <godot_cpp/variant/variant.hpp>
+
+#include <Ultralight/Ultralight.h>
+#include <AppCore/JSHelpers.h>
+
+using namespace ultralight;
+using namespace godot;
+using namespace std;
+
+typedef vector<shared_ptr<Variant>> VariantRefrences;
+
+class Convert {
+private:
+    static shared_ptr<VariantRefrences> variant_refrences;
+public:
+    static Variant ToVariant(JSContextRef context, JSValueRef value);
+    static JSValueRef ToJSValue(JSContextRef context, Variant variant);
+    static JSValueRef CallableCallback(JSContextRef ctx, JSObjectRef function, JSObjectRef thisObject, size_t argumentCount, const JSValueRef arguments[], JSValueRef* exception);
+};
+
+#endif
